Disable non-protobuf session test
It causes the error "k:0 is both fed and fetched". According to TF
rules, a mere constant by itself won't do. We'd need to add an
identity operation to the graph to be able to run it properly.

The graph generated from Python and loaded from the resulting Protocol
Buffers file seems to work with only one constant, I don't know what's
up with that.