Make TF variable initialization work
This was puzzling me for a while - I couldn't get a hold of
tf.global_variables_initializer() from the Java/Clojure side. Turns
out it was easy: we just need to use tf.variables_initializer()
instead. It takes an optional name= argument that we can use to give
the initializer a name that allows us to grasp at it using the Java
API.